+++ Engine/src/Renderer/Mesh.h
@@ -6,7 +6,7 @@ class Mesh {
 public:
 	Mesh(GLenum prim, unsigned int vert_col);
 	~Mesh();
-	void SetVertexContent(const std::vector<double>& vertices, unsigned int index);
+	void SetVertexContent(const std::vector<float>& vertices, unsigned int index);
 	void Draw();
 private:
 	GLuint vao;
